requirements.txt
gtfs-realtime-bindings==1.0.0
protobuf==5.26.1
PyYAML==6.0.1
pydantic==2.7.1
requests==2.31.0